Can now use -a to output metadata for pdfs.
Man, parsing pdfs is an enormous pain!  Finally had to use
fstat() to get the file size and parse it byte by byte,
because the pdf standard allows *multiple "%%EOF" strings in
a file*.  Not just multiple zero-bytes, which is bad enough;
we can get around that just by looking for the magic string
"%%EOF".  No, multiple "%%EOF" strings.  The insanity of
this is hard to describe.