Now I use pdfinfo to get pdf metadata.
Much nicer; actually works, even on the encrypted files,
which otherwise produce a bunch of malarky.  I did have to
extend the memory reserved for the field names due to this.
TODO:  check for the presence of pdfinfo and unzip, so that
I can print a warning if they're not present; look into
scrolling the detail view if the information is too large.