Added metadata extraction for *lots* of types.
Thanks to GNU libextractor, it was pretty straightforward.
It *doesn't* do pdf and epub, so that effort wasn't wasted.
It appears that it works, at least for ogv.  Also, changed
"OGT" and "OGV" to "OGV" and "OGG," which I think is
correct.  Finally, lengthened the pad for detail view,
because these media files have a *lot* of metadata.