Fixed a bug affecting detail view and opening.
Fields that started with a number were causing some really
bizarre detail view; so, e.g., record 1024 was showing
details from, and opening files from, the first record that
contained the number "1024" at the beginning of a field.
Field starts now have to both atoi() properly *and* include
the string "%%", so now it works.