Kept working on the statistics.
I can calculate the number of each tag; now I'd like to
calculate the number of unique values for each tag, which is
quite a bit harder.  I could just loop through the whole of
it for each tag and seek unique values, but that seems like
it would run *very* long.  Better would be to store values
I've encountered and then add one to a counter when the
value is unique; but that's tough, too, since I'd have to
dynamically allocate the structs.